Fixing e_factor column type

+++ database/migrations/2024_10_25_114206_create_lexemes_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('text');
             $table->string('meaning');
             $table->string('language');
-            $table->unsignedTinyInteger('e_factor')->default(2.5);
+            $table->float('e_factor')->default(2.5);
             $table->timestamps();
         });
     }
